move LessonCard files to new folder
