+++ components/LessonCard/LessonCard.tsx
@@ -1,11 +1,11 @@
 import { useQuery } from '@apollo/client'
 import React from 'react'
 import { CheckCircle } from 'react-feather'
-import GET_SUBMISSIONS from '../graphql/queries/getSubmissions'
-import NavLink from './NavLink'
+import GET_SUBMISSIONS from '../../graphql/queries/getSubmissions'
+import NavLink from '../NavLink'
 import Image from 'next/image'
-import styles from '../scss/lessonCard.module.scss'
-import { Submission, SubmissionStatus } from '../graphql'
+import styles from './lessonCard.module.scss'
+import { Submission, SubmissionStatus } from '../../graphql'
 import Link from 'next/link'
 import { get } from 'lodash'
