+++ components/LessonCard/lessonCard.module.scss
@@ -1,4 +1,4 @@
-@use './_variables.module';
+@use '../../scss/variables.module';
 
 .lesson-card__container_inprogress {
   border-width: 2px;
